requirements.txt
Flask==2.0.2
requests==2.26.0
SQLAlchemy==1.4.25
gunicorn==20.1.0
pymongo==4.11
python-dotenv==0.19.1
Pillow==8.4.0
Werkzeug==2.0.2
requests==2.26.0
python-dotenv==0.19.1
Flask-SocketIO==5.5.1
celery==5.4.0
redis==5.2.1